admin: login entrypoint minor refactoring
